test(core): split plugin onload loader tests
